Add ceph installation readme
This allow an installation of ceph to be used with BGS.
It basically needs 4 machines:
 * 1 ceph admin
 * 3 ceph nodes

The machines needs to be installed with Ubuntu 14.04 priorily with a mecanism
to be defined (e.g. PXE boot with fuel/cobler/maas).

The machines are connected with at least two networks interfaces:
 * 1 network for storage
 * 1 network for admin

This is only a documentation that would be integrated into python/bash scripts
to automate the installation when needed/ready.

Change-Id: I38656bcc21fa43d6cfeac9108f5804ce03ed960f
Signed-off-by: Arnaud Morin <arnaud1.morin@orange.com>