Update of Fuel based build system to Fuel 6.0.1
Rebase of the build system to build Fuel from the stable/6.0 branch
which currently creates Fuel 6.0.1. The Fuel baseline thus becomes:

	- OpenStack Juno
	- Ubuntu 12.04.4
	- CentOS 6.5

An additional patch is added to the Fuel build system to enable
a newer version of Fuel to be built under Docker (as the Ubuntu
sandbox chroot is installing atd, where the post install expects
upstart to be running which is not the case in the build container).

Rebase of patches on top of Fuel to support 6.0.1.

Change-Id: I2e884d27e1533be8d60c071a7ea9db998a4f5fea
Signed-off-by: Stefan K. Berg <stefan.k.berg@ericsson.com>