Initial setup of "common" directory tree
Complement the tree structure that was created by Joseph with
a directory tree that is to hold files which are independent
from a particular installer approach and as such are common
to the OPNFV install and maintenance process:

genesis
 └── common
      ├── manifests
      └── tools

Directories contain README files that describe their use.
Cleaned up trailing whitespaces.

Change-Id: Icad2a9e9e64764d8d495f779d73176a36208062e
Signed-off-by: Frank Brockners <fbrockne@cisco.com>