Adds deploy.sh for Foreman/QuickStack PATCHSET2: Fixes whitespace issue and adds default_gw substitution in ksgen settings file
Creates a Vagrant VM provisioned with VirtualBox.  The Vagrantfile
config is modified at runtime with dynamic support for baremetal network
setup.  IP addresses and other network info are gathered at runtime and
passed to the .yaml file which is fed into Khaleesi to install
Foreman/QuickStack and provision hosts.  For the script usage it is
recommended to include the following inside of the yaml file set:
 - mac_address: "01:23:45:67:89:AB" <-Mac Address of admin interface
 - bmc_ip: 10.4.17.2
 - bmc_mac: "01:23:45:67:88:AB"
 - bmc_user: root
 - bmc_pass: root

JIRA: BGS-31

Change-Id: I1aa9352d2f5965befabf86119c0c9c5edb82f329
Signed-off-by: Tim Rozet <trozet@redhat.com>