Adds support for subnets other than /24 and fixes issue where default_gw was set in the wrong place PATCHSET2: Fixes improper spacing in network_type for .yaml settings
Correct subnet is now detected . This was needed for Intel Pod1, but can
be used in other environments.  Vagrantfile is now updated to add
this correct netmask, and variables are updated with correct netmask as
well.  default_gw was being set before the base_config was copied, so we
were losing the value of the variable.  Moved this statement to be after
base_config.

JIRA: BGS-31

Change-Id: I36c69abf5e687fe19a92a9659b2069e1f32275a8
Signed-off-by: Tim Rozet <trozet@redhat.com>