Create Compass build script.
Compass is a open source platform for deploying distributed systems,
including OpenStack. It has a plugin structure for hardware discovery,
os provisioning and package installation and configuration.
As for OPNFV BGS experiment, we want to deploy it in automatic mechanism,
so in this script we build the ISO including compass and other packages
that could be used in later deployment.

JIRA: BGS-40

Change-Id: Ie08ef0911ff0d8f23c2d18d929b34c204f1a7536
Signed-off-by: chigang <chigang@huawei.com>