Auto deploy prototype #2
This is an evolvement of the libvirt deployer to make it fully
general, and it also add the concept of Deploy Environment
Adapter (dea), Deploy Hardware Adapter (dha) and a hardware
adapter plugin.

See the README.rst file for an introduction of how to start
experimenting with this concept in a libvirt setting.

Highlights:

- Can "xerox" an existing Fuel deploy
- Separates configuration for the deployment (DEA) and
  the configuration for the hardware (DHA)
- Introduces an API to be implemented by hardware adapters
  (a libvirt example adapter is included)
- Provides a verification tool "verify_dha.sh" to validate
  DHA adapters against the DHA API.

See the TODO.txt file for things that can be worked on.

JIRA

Change-Id: I98ff665e6b63c6c3bc42bb6b65c1d2151359e374
Signed-off-by: Stefan K. Berg <stefan.k.berg@ericsson.com>