Deployment corrections for Fuel
The most important change is the introduction of support for a selectable
temporary directory for ISO manipulations by setting the TMPDIR variable.

Invoke deploy.sh as "sudo TMPDIR=/my/tmp/dir ./deploy.sh ..." to make
use of this, otherwise the temporary directory defaults to that of the
user running the script.

Other improvements:

- Added error check for when the patch-iso.sh gets an error from
  copying data with cpio (potentially indicating out of disk space)
- Corrected typo in deploy.sh: error-exit -> error_exit

JIRA

Change-Id: I74966c3fe2bee0ac1923cfa9bd9ebfe1ca73a9da
Signed-off-by: Stefan K. Berg <stefan.k.berg@ericsson.com>