Adds external network support for OPNFV
Deploy will now detect and set extra parameters in the ksgen settings
file needed in order to configure an external public network for
Neutron. The new parameters required are as follows:
 -  external_network_flag: "true"
 -  public_gateway:
 -  public_dns:
 -  public_network:
 -  public_subnet:
 -  public_allocation_start:
 -  public_allocation_end:
 -  deployment_type:

public_* parameters are required to create an external network when
external_network_flag is "true".  deployment_type (three_network or
multi_network) is now passed to Foreman.  Foreman decides on this value
whether to configure 3 interfaces as dhcp+onboot or 4.

Change-Id: Ia8af04e82f3854a36adc563c1855d451ddfa9c17
JIRA: BGS-13
Signed-off-by: Tim Rozet <trozet@redhat.com>