Adding in support for Lithium container.
Both will be present on the controller nodes and can be started/switches on and off at will. Networking scripts are not updated - will be deprecated in
favour of plugin method of enabling / setting up VXLAN / OVSDB integration

pulling from the latest nightly artifact at time of commit - will work on automated pulling each night to see if that makes sense.

Change-Id: Ie911cdf61cd97a99b975c30c55c664daf70eb3ee
JIRA: 0
Signed-off-by: Dan Smith <daniel.smith@ericsson.com>