Fixing verification of vbox drivers JIRA: BGS-74
With the foreman install the vboxdrv module is not loaded
then the deploy.sh script attempts to load the driver by
running ‘/etc/init.d/vboxdrv setup’
and checks the return code for success.

However, /etc/init.d/vboxdrv will return 0 as long a valid parameter was passed on the command line. In the case of a
failure, a failure message will be logged and the return code
will still be 0.

For instance if the kvm module is already loaded, the
VirtualBox driver will never install.

deploy.sh will now check for the VirtualBox kernel module
with lsmod after the setup script is installed.

The deploy.sh will now exit if the vboxdrv is not loaded.

Change-Id: I702819cbf28afb08e0035e08918390af85c07674
Signed-off-by: randyl <r.levensalor@cablelabs.com>