Fixes networking related constraints for Virtual deployments
Patchset fixes the following:
 - Virtual deployment no longer requires 3 networks, does a lookup of
the public interface on the server to bridge to, and then uses internal
VirtualBox networks for the other admin and private networks
 - enable_virtual_dhcp option now allows enabling of Foreman DHCP
server.  This new default behavior is to use static IP address in
virtual deployments.  This is more useful in environments where
disabling DHCP on the public network is not an option.
 - static_ip_range argument used to define which IP block to use out of
the public network for the deployment. Valid only in virtual deployment.
 - puppet verification for virtual deployments added
 - horizon and foreman URLs will now be printed at the end of deployment
 - fixes and updates to non-HA deployments
 - ping_site argument now allows users to specify the site to verify VM
internet connectivity.  Default is google.com which is unreachable in
China

JIRA: BGS-73

Change-Id: I63b4c91477591d2f9436b5e6f59a2f2ee021d7d7
Signed-off-by: Tim Rozet <trozet@redhat.com>