Add a 1 minute pause after ansible playbook
For the LF pods, it seems that the ansible reload playbook is necessary,
in our pod, we also need to wait 1 minute before trying to run shell
provisionner in Vagrant VM.
This patch adds a 1 minute pause into the ansible playground in order to
be sure that the machine will be accessible.
This avoid "SSH connection unexpectedly closed" issue.

Change-Id: Iab2b47decb0120e8359a175b1f16ccbd1036e91f
JIRA: BGS-85
Signed-off-by: arnaudmorin <arnaud.morin@gmail.com>