Take care of the subnet on public subnet
This patch will take care of the subnet on public range. Instead of
having a stoned /24, we can use now use at least /27 subnet (because
foreman will try to use 20 IPs for public floating ip pool)
This is not the best way to do that, but it's better than the current
way.

It also add a parameter to set the number of floating IP we want to use
from the public subnet in provider network.

Change-Id: I467f2a4098d2da3c6f666453cead64e18d0c655c
JIRA: BGS-75
Signed-off-by: arnaudmorin <arnaud.morin@gmail.com>