Fixed public IP allocation JIRA: BGS-89
For bare metal deployments, the list IP of the last IP of the subnet
and 20 IPs before are reservered for neutron public IPs.

A new function to subtract a number from an IP was added.

Really fixed the logic for skipping 10 IP for VIPs was flipped to treat
no_dhcp with static IPs as a special case.

Removed a false print statement that the foreman VM is running.
Consolidated two identical print messages with the public IP range.

Change-Id: I189b9440dafd98e49d9a147515810e44b97ab256
Signed-off-by: randyl <r.levensalor@cablelabs.com>