Fixes bug with interface detection
This patch addresses the detection of the physical NICs to bridge to.
There was a bug where the regexp would ignore some NICs on accident, and
the purpose of the expression was to ignore virtual NICs.  Rather than
ignore NICs by their name, this patch ensures that NICs are ignored if
they are actually virtual types.

Also fixes syntax issue with checking for nic_arg_flag.

JIRA: APEX-11

Change-Id: Ibd965c454faab1d751e3ce0b260eabc7fbe7c720
Signed-off-by: Tim Rozet <trozet@redhat.com>