Adds ability to specify NICs to bridge on the jumphost
This patch adds new functionality to be able to specify the physical
interfaces on the jumphost to bridge to (rather than figured out
dynamically).  The patch introduces new args for each nic:
 - admin_nic
 - private_nic
 - public_nic
 - storage_nic

The public_nic is the only arg to be used when using -virtual.  In
baremetal: admin_nic, private_nic, and public_nic must either be all or
none specified.  The patch also cleans up virtual node logic, which was
working, but had unnecessary logic.

JIRA: APEX-4

Change-Id: Ic0a2e1d8ce2b7464441e3e9041faf123b86ca7e6
Signed-off-by: Tim Rozet <trozet@redhat.com>