Fixes issue with deploy copying from /tmp and clean bailing early
deploy.sh was failing in user deployments because the necessary puppet
modules and vagrant files were being copied accidentally from /tmp and
not the script directory.  This patch fixes that issue.  clean.sh was
failing checks to see if vagrant or virtualbox was still running, due to
a bug with checking "ps" and using grep.  This patch resolves that.

JIRA: APEX-16,APEX-17

Change-Id: I1faa6fc134c0308acb2e7b14be30f7cd3c99d109
Signed-off-by: Tim Rozet <trozet@redhat.com>