Fixes horizon IP URL for non-HA deployments
The printed horizon URL at the end of a successful deployment was
incorrect for non-HA deployments.  The private IP was being printed
which is not publicly accessible.  This patch fixes the URL printed to
be the public IP of the control node - the correct URL.

JIRA: APEX-12

Change-Id: Ia2f594e0cf839a9100c3bf5715a38fe4fd243bfe
Signed-off-by: Tim Rozet <trozet@redhat.com>
(cherry picked from commit db4a9ac5492659cd435b4e4e84868510259c6b29)