Corrects the default route on controllers
BGS-84

After the changes to br-ex on the control nodes.  The default route
was no longer correct on the control nodes in bare metal deployments.

Added an admin_network paramater for barmetal deployments.

Added logic on br-ex in to set DEROUTE=no on the admin network
in the extern_net_presetup.

Tested on HA physical deployment.

Change-Id: Ic4b1ef521da2c55590990411f3f2444d4c94b019
Signed-off-by: randyl <r.levensalor@cablelabs.com>