Ubuntu package repositories added to Fuel ISO
Starting with Fuel 6.1, the Ubuntu repositories are no longer stored
on the Fuel ISO, but by default fetched from the Internet during
installation (the exact URI can be modified in the Fuel configuration).

In order to being able to handle a deployment scenario where the Fuel
node has no Internet connectivity, the OPNFV Fuel ISO is adding the
necessary repositories into /var/www/nailgun by fetching them during
the generation of the ISO.

The actual fetching is done by a Fuel provided package,
fuel-createmirror.deb, which can be found in /opt of the Fuel master
(but which is run outside Fuel in the context of the build process
when generating the OPNFV ISO).

An additional OPNFV post installation script is added as
/opt/opnfv/bootstrap/post.d/03_install_repo.sh
which will populate /var/www/nailgun with these repositories in the
late stage of the bootstrap_admin_node process.

Important!

In order to use these repositories For deployment, the default Fuel
repo configuration must be changed as below:

Repository "mos-updates"
 URI="deb http://10.20.0.2:8080/mos-ubuntu mos6.1-updates main restricted"
Repository "mos-security"
  URI="deb http://10.20.0.2:8080/mos-ubuntu mos6.1-security main restricted"
Repository "mos-holdback"
  URI="deb http://10.20.0.2:8080/mos-ubuntu mos6.1-holdback main restricted"

Repository "ubuntu"
  URI="deb http://10.20.0.2:8080/ubuntu-part trusty main"
Repository "ubuntu-security"
  URI="deb http://10.20.0.2:8080/ubuntu-part trusty main"
Repository "ubuntu-updates"
  URI="deb http://10.20.0.2:8080/ubuntu-part trusty main"

Change-Id: I25211bd3f277b560808fd838180aa52e278af404
Signed-off-by: Stefan K. Berg <stefan.k.berg@ericsson.com>