Ading CLI verification / working to documentation
Adding command line only verifciation for OpenStack for
environets without a web browswer as an alternatiative
method.

Minor changes for logging into the foremen VM that reflects
the locked down permission of the foreman VM directory.

Added BGS-84 as a fix to the release notes.

Added steps to workaorund the known neutron DHCP issue.

Change-Id: Icfea3415187de33be5aa44befe93c7249622227d
Signed-off-by: randyl <r.levensalor@cablelabs.com>